+++ pywr/_core.pyx
@@ -218,32 +218,24 @@ cdef class ScenarioIndex:
 
 
 cdef class Timestep:
-    def __init__(self, datetime, int index, double days):
-        self._datetime = pd.Timestamp(datetime)
-        self._index = index
-        self._days = days
-        tt = self.datetime.timetuple()
-        self.dayofyear = tt.tm_yday
-        self.day = tt.tm_mday
-        self.month = tt.tm_mon
-        self.year = tt.tm_year
+    def __init__(self, period, int index, double days):
+        self.period = period
+        self.index = index
+        if days <= 0:
+            raise ValueError("The days argument must be > 0.")
+        self.days = days
+        self.dayofyear = period.dayofyear
+        self.day = period.day
+        self.month = period.month
+        self.year = period.year
 
     property datetime:
         """Timestep representation as a `datetime.datetime` object"""
         def __get__(self, ):
-            return self._datetime
-
-    property index:
-        """The index of the timestep for use in arrays"""
-        def __get__(self, ):
-            return self._index
-
-    property days:
-        def __get__(self, ):
-            return self._days
+            return self.period.to_timestamp()
 
     def __repr__(self):
-        return "<Timestep date=\"{}\">".format(self._datetime.strftime("%Y-%m-%d"))
+        return "<Timestep date=\"{}\">".format(self.period.strftime("%Y-%m-%d"))
 
 cdef class Domain:
     """ Domain class which all Node objects must have. """
@@ -387,6 +379,7 @@ cdef class AbstractNode:
         cdef int i
         for i in range(self._flow.shape[0]):
             self._flow[i] = 0.0
+            self._prev_flow[i] = 0.0
 
     cpdef before(self, Timestep ts):
         """Called at the beginning of the timestep"""
@@ -898,7 +891,14 @@ cdef class Storage(AbstractStorage):
         AbstractStorage.reset(self)
         self._reset_storage_only()
 
-    cpdef _reset_storage_only(self):
+    cpdef _reset_storage_only(self, bint use_initial_volume=True):
+        """Reset the current volume of the storage node.
+
+        Parameters
+        ==========
+        use_initial_volume : bool (default: True)
+            Reset the volume to the initial volume of the storage node. If false the volume is reset to max_volume.
+        """
         cdef int i
         cdef double mxv = self._max_volume
         cdef ScenarioIndex si
@@ -929,12 +929,15 @@ cdef class Storage(AbstractStorage):
             if self._max_volume_param is not None:
                 mxv = self._max_volume_param.get_value(si)
 
-            if np.isfinite(self._initial_volume_pc):
-                self._volume[i] = self._initial_volume_pc * mxv
-            elif np.isfinite(self._initial_volume):
-                self._volume[i] = self._initial_volume
+            if use_initial_volume:
+                if np.isfinite(self._initial_volume_pc):
+                    self._volume[i] = self._initial_volume_pc * mxv
+                elif np.isfinite(self._initial_volume):
+                    self._volume[i] = self._initial_volume
+                else:
+                    raise RuntimeError('Initial volume must be set as either a percentage or absolute volume.')
             else:
-                raise RuntimeError('Initial volume must be set as either a percentage or absolute volume.')
+                self._volume[i] = mxv
 
             try:
                 self._current_pc[i] = self._volume[i] / mxv
@@ -948,7 +951,7 @@ cdef class Storage(AbstractStorage):
         cdef ScenarioIndex si
 
         for i, si in enumerate(self.model.scenarios.combinations):
-            self._volume[i] += self._flow[i]*ts._days
+            self._volume[i] += self._flow[i]*ts.days
             # Ensure variable maximum volume is taken in to account
 
             mxv = self.get_max_volume(si)
@@ -991,6 +994,7 @@ cdef class AggregatedStorage(AbstractStorage):
         cdef ScenarioIndex si
 
         for i, si in enumerate(self.model.scenarios.combinations):
+            mxv = 0.0
             for s in self._storage_nodes:
                 mxv += s.get_max_volume(si)
 
@@ -1013,7 +1017,7 @@ cdef class AggregatedStorage(AbstractStorage):
             for s in self._storage_nodes:
                 self._flow[i] += s._flow[i]
                 mxv += s.get_max_volume(si)
-            self._volume[i] += self._flow[i]*ts._days
+            self._volume[i] += self._flow[i]*ts.days
 
             # Ensure variable maximum volume is taken in to account
             try: