+++ pywr/_model.pyx
@@ -2,7 +2,6 @@ import os
 import pandas
 import json
 import networkx as nx
-from past.builtins import basestring
 import copy
 from packaging.version import parse as parse_version
 import warnings
@@ -26,7 +25,7 @@ from pywr._component cimport Component
 from pywr.nodes import Storage, AggregatedStorage, AggregatedNode, VirtualStorage
 from pywr._core import ScenarioCollection, Scenario
 from pywr._core cimport AbstractNode
-from pywr.parameters._parameters import load_dataframe
+from .dataframe_tools import load_dataframe
 from pywr.parameters._parameters import Parameter as BaseParameter
 from pywr.parameters._parameters cimport Parameter as BaseParameter
 from pywr.recorders import ParameterRecorder, IndexParameterRecorder, Recorder
@@ -113,6 +112,8 @@ class Model(object):
             key = list(kwargs.keys())[0]
             raise TypeError("'{}' is an invalid keyword argument for this function".format(key))
 
+        self._time_before = None
+        self._time_after = None
         self.reset()
 
     @property
@@ -207,29 +208,45 @@ class Model(object):
         """
         if "includes" in data:
             for filename in data["includes"]:
+                _, ext = os.path.splitext(filename)
                 if path is not None:
                     filename = os.path.join(os.path.dirname(path), filename)
-                with open(filename, "r") as f:
-                    try:
-                        include_data = json.loads(f.read())
-                    except ValueError as e:
-                        message = e.args[0]
-                        if path:
-                            e.args = ("{} [{}]".format(e.args[0], os.path.basename(filename)),)
-                        raise(e)
-                for key, value in include_data.items():
-                    if isinstance(value, list):
-                        try:
-                            data[key].extend(value)
-                        except KeyError:
-                            data[key] = value
-                    elif isinstance(value, dict):
-                        try:
-                            data[key].update(value)
-                        except KeyError:
-                            data[key] = value
-                    else:
-                        raise TypeError("Invalid type for key \"{}\" in include \"{}\".".format(key, path))
+
+                ext = ext.lower()
+                if ext == '.json':
+                    cls._load_json_include(data, filename)
+                elif ext == '.py':
+                    cls._load_py_include(filename)
+                else:
+                    raise NotImplementedError(f'Include file type "{ext}" not supported.')
+
+    @classmethod
+    def _load_py_include(cls, filename):
+        import runpy
+        runpy.run_path(filename)
+
+    @classmethod
+    def _load_json_include(cls, data, filename):
+        with open(filename, "r") as f:
+            try:
+                include_data = json.loads(f.read())
+            except ValueError as e:
+                message = e.args[0]
+                e.args = ("{} [{}]".format(e.args[0], os.path.basename(filename)),)
+                raise(e)
+        for key, value in include_data.items():
+            if isinstance(value, list):
+                try:
+                    data[key].extend(value)
+                except KeyError:
+                    data[key] = value
+            elif isinstance(value, dict):
+                try:
+                    data[key].update(value)
+                except KeyError:
+                    data[key] = value
+            else:
+                raise TypeError("Invalid type for key \"{}\" in include \"{}\".".format(key, filename))
         return None  # data modified in-place
 
     @classmethod
@@ -249,7 +266,7 @@ class Model(object):
             Name of the solver to use for the model. This overrides the solver
             section of the model document.
         """
-        if isinstance(data, basestring):
+        if isinstance(data, str):
             # argument is a filename
             logger.info('Loading model from file: "{}"'.format(path))
             path = data
@@ -297,7 +314,7 @@ class Model(object):
         else:
             start = pandas.to_datetime(timestepper_data['start'])
             end = pandas.to_datetime(timestepper_data['end'])
-            timestep = int(timestepper_data['timestep'])
+            timestep = timestepper_data['timestep']
 
         if model is None:
             model = cls(
@@ -550,7 +567,7 @@ class Model(object):
 
 
         """
-        if self.dirty:
+        if self.dirty or self.timestepper.dirty:
             self.setup()
         self.timestep = next(self.timestepper)
         return self._step()
@@ -573,9 +590,8 @@ class Model(object):
         t0 = time.time()
         timestep = None
         try:
-            if self.dirty:
+            if self.dirty or self.timestepper.dirty:
                 self.setup()
-                self.timestepper.reset()
             else:
                 self.reset()
             t1 = time.time()
@@ -602,6 +618,8 @@ class Model(object):
             num_scenarios=num_scenarios,
             timestep=timestep,
             time_taken=time_taken,
+            time_taken_before=self._time_before,
+            time_taken_after=self._time_after,
             time_taken_with_overhead=time_taken_with_overhead,
             speed=speed,
             solver_name=self.solver.name,
@@ -615,6 +633,7 @@ class Model(object):
         """Setup the model for the first time or if it has changed since
         last run."""
         logger.info('Setting up model ...')
+        self.timestepper.setup()
         self.scenarios.setup()
         length_changed = self.timestepper.reset()
         for node in self.graph.nodes():
@@ -645,6 +664,9 @@ class Model(object):
             component.reset()
 
         self.solver.reset()
+        # reset the timers
+        self._time_before = 0.0
+        self._time_after = 0.0
         logger.info('Reset complete!')
 
     def before(self):
@@ -660,6 +682,7 @@ class Model(object):
         cdef AbstractNode node
         cdef Component component
         cdef BaseParameter param
+        cdef double t0 = time.time()
         for node in self.graph.nodes():
             node.before(self.timestep)
         cdef list components = self.flatten_component_tree(rebuild=False)
@@ -669,15 +692,18 @@ class Model(object):
             if isinstance(component, BaseParameter):
                 param = component
                 param.calc_values(self.timestep)
+        self._time_before += time.time() - t0
 
     def after(self):
         cdef AbstractNode node
         cdef Component component
+        cdef double t0 = time.time()
         for node in self.graph.nodes():
             node.after(self.timestep)
         cdef list components = self.flatten_component_tree(rebuild=False)
         for component in components:
             component.after()
+        self._time_after += time.time() - t0
 
     def finish(self):
         for node in self.graph.nodes():
@@ -770,7 +796,7 @@ class NodeIterator(object):
 
     def __delitem__(self, key):
         """Remove a node from the graph"""
-        if isinstance(key, basestring):
+        if isinstance(key, str):
             node = self[key]
         else:
             node = key
@@ -799,6 +825,12 @@ class NodeIterator(object):
     def __len__(self):
         """Returns the number of nodes in the model"""
         return len(list(self._nodes()))
+
+    def __contains__(self, value):
+        for node in self._nodes():
+            if node.name == value or node == value:
+                return True
+        return False
 
     def __iter__(self):
         return self
@@ -859,17 +891,25 @@ class NamedIterator(object):
     def __iter__(self):
         return iter(self._objects)
 
+    def __contains__(self, value):
+        for obj in self._objects:
+            if obj.name == value or obj == value:
+                return True
+        return False
+
     def append(self, obj):
         # TODO: check for name collisions / duplication
         self._objects.append(obj)
 
 
 class ModelResult(object):
-    def __init__(self, num_scenarios, timestep, time_taken, time_taken_with_overhead, speed,
-                 solver_name, solver_stats, version):
+    def __init__(self, num_scenarios, timestep, time_taken, time_taken_before, time_taken_after, time_taken_with_overhead,
+                 speed, solver_name, solver_stats, version):
         self.timestep = timestep
         self.timesteps = timestep.index + 1
         self.time_taken = time_taken
+        self.time_taken_before = time_taken_before
+        self.time_taken_after = time_taken_after
         self.time_taken_with_overhead = time_taken_with_overhead
         self.speed = speed
         self.num_scenarios = num_scenarios