+++ tests/test_core.py
@@ -1,13 +1,12 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from __future__ import print_function
 import datetime
 import pytest
 from fixtures import *
 from helpers import *
 from pywr._core import Timestep, ScenarioIndex
-
+import pandas
 from pywr.core import *
 from pywr.domains.river import *
 from pywr.parameters import Parameter, ConstantParameter, DataFrameParameter, AggregatedParameter
@@ -209,8 +208,9 @@ def test_timeseries_excel(simple_linear_model, filename):
     ts = load_parameter(model, data)
 
     # model (intentionally not aligned)
-    model.timestepper.start = ts.dataframe.index[0] + 5
-    model.timestepper.end = ts.dataframe.index[-1] - 12
+    index = ts.dataframe.index
+    model.timestepper.start = index[0] + (5 * index.freq)
+    model.timestepper.end = index[-1] - (12 * index.freq)
 
     # need to assign parameter for it's setup method to be called
     model.nodes["Input"].max_flow = ts
@@ -427,10 +427,8 @@ def test_storage_initial_volume_pc():
     np.testing.assert_allclose(storage.volume, 20.0)
 
 
-def test_storage_max_volume_param_raises():
-    """Test a that an max_volume with a Parameter that has children.
-    
-    Only some aggregated style parameters should work here.
+def test_storage_max_volume_nested_param():
+    """Test that a max_volume can be used with a Parameter with children.
     """
 
     model = Model(
@@ -452,8 +450,7 @@ def test_storage_max_volume_param_raises():
 
 
 def test_storage_max_volume_param_raises():
-    """Test a that an max_volume with a Parameter that has children is an error
-
+    """Test that a max_volume can not be used with 2 levels of child parameters.
     """
 
     model = Model(
@@ -508,6 +505,15 @@ def test_json_include():
     supply1 = model.nodes["supply1"]
     supply2 = model.nodes["supply2"]
     assert(isinstance(supply2.max_flow, ConstantParameter))
+
+
+def test_py_include():
+    """Test include in Python document"""
+    filename = os.path.join(TEST_FOLDER, "models", "python_include.json")
+    model = Model.load(filename)
+
+    model.run()
+
 
 def test_json_min_version():
     """Test warning is raised if document minimum version is more than we have"""
@@ -577,3 +583,19 @@ def test_variable_load():
     assert sorted([c.is_objective for c in model.objectives]) == ["minimise", ]
     assert sorted([c.name for c in model.constraints]) == ["min_volume", ]
 
+
+def test_delta_greater_than_zero_days(simple_linear_model):
+    """Test trying to use zero length timestep raises an error."""
+    model = simple_linear_model
+    model.timestepper.delta = 0
+
+    with pytest.raises(ValueError):
+        model.run()
+
+
+def test_timestep_greater_than_zero_days():
+    """Test trying to create zero length Timestep."""
+
+    with pytest.raises(ValueError):
+        # Test setting days <= 0 raises an error
+        Timestep(pandas.Period('2019-01-01', freq='D'), 0, 0)