+++ travis/build-wheels.sh
@@ -13,11 +13,11 @@ export PYWR_BUILD_LPSOLVE=true
 # Install the build dependencies of the project. If some dependencies contain
 # compiled extensions and are not provided as pre-built wheel packages,
 # pip will build them from source matching the target Python version and architecture
-pip install cython packaging numpy jupyter pytest pytest-cov wheel future setuptools_scm
-# Install run-time packages
-pip install platypus-opt inspyred pygmo
+pip install cython packaging numpy jupyter pytest pytest-cov wheel setuptools_scm 'coverage<5.0'
+# Install run-time packages; install is allowed to fail for these as they are optional.
+pip install platypus-opt inspyred pygmo || true
 
-if [[ "${PYWR_BUILD_TRACE}" -eq "true" ]]; then
+if [[ "${PYWR_BUILD_TRACE}" == "true" ]]; then
     # For tracing we need to build in-place
     python setup.py build_ext -i develop